refactor(trusty-search): split core/indexer/{ingest,search} under 500 SLOC (#607 PR-E)
Split the two largest monolith files in core/indexer to satisfy the 500-line
hard cap enforced by check_line_cap.sh:

ingest.rs (1214 lines) → ingest/{mod,embed,commit}.rs
  - ingest/mod.rs  (498 lines): all parse+embed+batch orchestration methods
  - ingest/embed.rs (246 lines): resolve_embed_inflight, embed_chunks_in_batches
  - ingest/commit.rs (356 lines): commit_parsed_batch, commit_vectors_batch,
    commit_bm25_batch, commit_corpus, commit_entities

search.rs (1109 lines) → search/{mod,kg,lanes,materialize}.rs
  - search/mod.rs  (465 lines): search(), apply_archive_downrank,
    apply_score_adjustments, inject_entity_exact_match, apply_mmr_rerank,
    merge_grep_lane, resolve_branch_set (orchestration only)
  - search/kg.rs    (161 lines): kg_expand, expand_with_kg, expand_with_kg_for_test
  - search/lanes.rs (238 lines): fetch_chunks_for_ids, get_embedding, embed_text,
    embed_query, bm25_search, grep_fallback_search, vector_search, edge_kinds_for_intent
  - search/materialize.rs (72 lines): materialize_search_results

Also splits out helpers.rs, persist_hnsw.rs, docs_penalty_tests.rs from the
original mod.rs. All 829 non-ignored tests pass; clippy clean; fmt clean;
line-cap check passes (164 allowlisted, 0 violations).
