+++ .line-cap-allowlist.tsv
@@ -20,7 +20,7 @@ crates/trusty-analyze/src/lang/adapters/ruby.rs	674
 crates/trusty-analyze/src/lang/adapters/rust.rs	530
 crates/trusty-analyze/src/lang/adapters/scala.rs	583
 crates/trusty-analyze/src/lang/adapters/swift.rs	531
-crates/trusty-analyze/src/lang/adapters/typescript.rs	680
+crates/trusty-analyze/src/lang/adapters/typescript.rs	674
 crates/trusty-analyze/src/main.rs	870
 crates/trusty-analyze/src/mcp/mod.rs	1157
 crates/trusty-analyze/tests/stdio_harness.rs	639
@@ -136,16 +136,12 @@ crates/trusty-review/src/pipeline/verify_tests.rs	556
 crates/trusty-search/src/commands/doctor_checks.rs	612
 crates/trusty-search/src/commands/integrate.rs	688
 crates/trusty-search/src/commands/reindex_engine.rs	1862
-crates/trusty-search/src/commands/reindex_ui.rs	1458
-crates/trusty-search/src/commands/start.rs	1893
+crates/trusty-search/src/commands/reindex_ui.rs	1457
+crates/trusty-search/src/commands/start.rs	1889
 crates/trusty-search/src/core/chunker/mod.rs	1883
 crates/trusty-search/src/core/classifier.rs	1032
 crates/trusty-search/src/core/corpus.rs	1685
-crates/trusty-search/src/core/indexer/docs_penalty.rs	543
-crates/trusty-search/src/core/indexer/ingest.rs	1214
-crates/trusty-search/src/core/indexer/mod.rs	1307
-crates/trusty-search/src/core/indexer/persist.rs	740
-crates/trusty-search/src/core/indexer/search.rs	1109
+crates/trusty-search/src/core/indexer/mod.rs	656
 crates/trusty-search/src/core/indexer/tests.rs	3426
 crates/trusty-search/src/core/memory_policy.rs	1264
 crates/trusty-search/src/core/migration/mod.rs	649
@@ -157,8 +153,8 @@ crates/trusty-search/src/core/symbol_graph.rs	1667
 crates/trusty-search/src/main.rs	1335
 crates/trusty-search/src/mcp/tools.rs	2187
 crates/trusty-search/src/service/call_chain.rs	954
-crates/trusty-search/src/service/daemon.rs	701
-crates/trusty-search/src/service/embedder_supervisor.rs	1020
+crates/trusty-search/src/service/daemon.rs	697
+crates/trusty-search/src/service/embedder_supervisor.rs	1018
 crates/trusty-search/src/service/grep.rs	827
 crates/trusty-search/src/service/persistence.rs	1017
 crates/trusty-search/src/service/reindex/mod.rs	2855
